them tinh nang cho payment

+++ EXE201/EXE201.Data/DTOs/CreatePaymentRequest.cs
@@ -8,7 +8,7 @@ namespace EXE201.Data.DTOs
 {
     public class CreatePaymentRequest
     {
-        public long OrderId { get; set; }
+        public int OrderId { get; set; }
         //public int Amount { get; set; }
         //public string Description { get; set; } = string.Empty;
     }
